feat(ui): separate Tailwind CSS plugin default export vs module exports (leading to often failures when mixin exports)
