+++ packages/ui/src/plugin/tailwindcss/config.ts
@@ -1,0 +1,8 @@
+import type { Config } from "tailwindcss";
+import { theme } from "./theme";
+
+export const config: Partial<Config> = {
+  theme: {
+    extend: theme,
+  },
+};
