+++ packages/ui/src/plugin/tailwindcss/index.ts
@@ -1,0 +1,9 @@
+import plugin from "tailwindcss/plugin";
+import { config } from "./config";
+
+export default plugin(
+  // plugin
+  () => {},
+  // config
+  config,
+);
